third_party/WebKit/LayoutTests/platform/win/webaudio/Oscillator/osc-square-sweep-snr-expected.txt

CONSOLE ERROR: Uncaught (in promise) loadFile: Request failed when loading ../Oscillator/oscillator-square-expected.wav. ()
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = undefined
PASS # AUDIT TASK RUNNER STARTED. 
PASS > [test]  
Harness: the test ran to completion.
